script: Add bash script to compute resource value from an Eve ENV
Environmental variables are not expanded when using `bash -c` so
we cannot use this in our `eve.yaml`.

We need to find a way to substitute these variables, so we have
added this bash script to make use of `envs` passed to the script directly.

This script returns the corresponding K8s value returned from a
K8s resource.